feat(assets): refactor assets structure + webpack encore config
