--- web/app/themes/adeliom/assets/components/admin/index.pcss
@@ -1,4 +1,0 @@
-body {
-    background: red;
-    background-image: url("../../images/logo_adeliom.svg");
-}
