--- web/app/themes/adeliom/assets/components/admin/index.ts
@@ -1,2 +1,0 @@
-import "@tailwind";
-import './index.pcss'
