--- web/app/themes/adeliom/assets/components/editor/index.ts
@@ -1,1 +1,0 @@
-import "@tailwind";
