--- web/app/themes/adeliom/assets/components/global/index.ts
@@ -1,2 +1,0 @@
-import "@tailwind";
-import "@components/navigations";
