--- web/app/themes/adeliom/assets/components/navigations/index.ts
@@ -1,1 +1,0 @@
-import "@components/navigations/menu";
