+++ web/app/themes/adeliom/assets/scripts/components/lightbox.ts
