+++ web/app/themes/adeliom/assets/scripts/components/modals.ts
@@ -1,4 +1,3 @@
-import "./index.pcss";
 import MicroModal from "micromodal";
 
 MicroModal.init({
