+++ web/app/themes/adeliom/assets/scripts/global.ts
@@ -1,0 +1,1 @@
+import "@scripts/navigations/menu";
