+++ web/app/themes/adeliom/assets/scripts/types/index.d.ts
