+++ web/app/themes/adeliom/assets/styles/admin.pcss
@@ -1,0 +1,3 @@
+.admin {
+    display: block;
+}
