+++ web/app/themes/adeliom/assets/styles/editor.pcss
@@ -1,0 +1,3 @@
+.editor {
+    display: block;
+}
