+++ web/app/themes/adeliom/assets/styles/global.pcss
@@ -1,0 +1,1 @@
+@import "tailwind/tailwind.pcss";
