+++ web/app/themes/adeliom/assets/styles/tailwind/base/base.pcss
@@ -1,0 +1,12 @@
+/**
+ * Custom styles to immediately follow Tailwind’s `base` layer
+ *
+ * “Add things like base typography styles [or] opinionated global resets.”
+ *
+ * — from https://tailwindcss.com/docs/plugins#adding-base-styles
+ */
+@layer base {
+    body {
+        overflow-x: hidden;
+    }
+}
