+++ web/app/themes/adeliom/assets/styles/tailwind/components/components.pcss
@@ -1,0 +1,5 @@
+@layer components {
+    .my-component {
+        display: block;
+    }
+}
