+++ web/app/themes/adeliom/assets/styles/tailwind/tailwind.pcss
@@ -1,0 +1,9 @@
+@import "tailwindcss/base";
+@import "base/base.pcss";
+
+@import "tailwindcss/components";
+@import "components/components.pcss";
+@import "components/modals.pcss";
+
+@import "tailwindcss/utilities";
+@import "utilities/utilities.pcss";
