+++ web/app/themes/adeliom/assets/styles/tailwind/utilities/utilities.pcss
@@ -1,0 +1,32 @@
+@layer utilities {
+    /************************************************
+                POSITION
+ *************************************************/
+    .absolute-center {
+        @apply absolute inset-y-2/4 inset-x-2/4 -translate-x-1/2 -translate-y-1/2;
+    }
+
+    .center-top {
+        @apply absolute top-1/2 -translate-y-1/2;
+    }
+
+    .center-both {
+        @apply absolute top-1/2 left-1/2 -translate-y-1/2 -translate-x-1/2;
+    }
+
+    .absolute-full {
+        @apply absolute inset-0 h-full w-full;
+    }
+
+    .fixed-full {
+        @apply fixed inset-0 h-full w-full;
+    }
+
+    .cover-full {
+        @apply absolute inset-0 h-full w-full object-cover;
+    }
+
+    .contain-full {
+        @apply absolute inset-0 h-full w-full object-contain;
+    }
+}
