--- web/app/themes/adeliom/assets/tailwind/components/_btn.pcss
@@ -1,3 +1,0 @@
-.btn {
-    @apply inline-block;
-}
