--- web/app/themes/adeliom/assets/tailwind/index.pcss
@@ -1,3 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
