--- web/app/themes/adeliom/assets/tailwind/index.ts
@@ -1,1 +1,0 @@
-import "./index.pcss";
